ct/reconciler: Move the reconciler to top-level CT
It's between level one and level zero.
