+++ src/v/cloud_topics/reconciler/range_batch_consumer.h
