+++ src/v/cloud_topics/reconciler/reconciler.h
@@ -14,7 +14,7 @@
 #include "base/seastarx.h"
 #include "cloud_io/remote.h"
 #include "cloud_storage_clients/types.h"
-#include "cloud_topics/level_zero/reconciler/range_batch_consumer.h"
+#include "cloud_topics/reconciler/range_batch_consumer.h"
 #include "cluster/notification.h"
 #include "cluster/partition.h"
 #include "cluster/partition_manager.h"
